backend/requirements.txt
asgiref==3.6.0
astropy==5.1.1
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
cloud-init==22.4.2
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
entrypoints==0.3
ffmpeg==1.4
filelock==3.9.0
Flask==2.2.3
gekko==1.0.5
gmplot==1.4.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
MarkupSafe==2.1.2
more-itertools==4.2.0
mpmath==1.2.1
netifaces==0.10.4
numpy==1.23.4
oauthlib==3.1.0
packaging==21.3
pexpect==4.6.0
pipenv==2023.2.18
platformdirs==3.1.0
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
pyerfa==2.0.0.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyInputPlus==0.2.12
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
PySimpleGUI==4.60.4
PySimpleValidate==0.2.12
python-apt==2.0.0+ubuntu0.20.4.8
python-debian===0.1.36ubuntu1
python-decouple==3.8
pytube==12.1.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.9.3
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.4
sqlparse==0.4.3
ssh-import-id==5.10
stdiomask==0.0.6
sympy==1.11.1
systemd-python==234
Twisted==18.9.0
ufw==0.36
urllib3==1.25.8
virtualenv==20.20.0
virtualenv-clone==0.5.7
wadllib==1.3.3
Werkzeug==2.2.3
zipp==1.0.0
zope.interface==4.7.1
